pyproject.toml
[project]
name = "imgbed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arq>=0.27.0",
    "fastapi>=0.135.3",
    "pillow>=12.2.0",
    "psycopg[binary]>=3.3.3",
    "python-multipart>=0.0.24",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]
